History mockup delete
