+++ src/pages/History/History.jsx
@@ -19,8 +19,8 @@ import { createTheme, ThemeProvider } from '@mui/material/styles';
 
 
 import * as React from 'react';
-import AspectRatio from '@mui/joy/AspectRatio';
-import CardOverflow from '@mui/joy/CardOverflow';
+// import AspectRatio from '@mui/joy/AspectRatio';
+// import CardOverflow from '@mui/joy/CardOverflow';
 // import Button from '@mui/joy/Button';
 // import Card from '@mui/joy/Card';
 // import CardContent from '@mui/joy/CardContent';
@@ -30,6 +30,11 @@ import CardOverflow from '@mui/joy/CardOverflow';
 import API from '../../api/axios';
 // import UserCreate from './Create';
 
+//Component
+import EditExercise from '../Exercise/EditExercise/EditExercise.jsx';
+
+//Image
+import createExercise from "../../components/UploadImage/UploadImage.jsx";
 // codeHistory
 function Copyright() {
   return (
@@ -47,11 +52,16 @@ function Copyright() {
 // TODO remove, this demo shouldn't need to reset the theme.
 const defaultTheme = createTheme();
 
-const Historyv1 = () =>  {
+const History = () =>  {
+  const [reload, setReload] = useState(false);
   const [exerciseActivities, setExerciseActivities] = useState([]);
   const exerciseActivitiesRoute = "api/exercise-activities";
   const user_id = "65c8f4fc9aa2f67b7fbd27a7";
+  
+  // const id = "65c9db8f4f0314251c487a0a";
+
   const token = localStorage.getItem('token');  // เก็บ token  login ค้างไว้
+
 
   const headers = {
       'Authorization': `Bearer ${token}`   // ขอ Token ก่อน
@@ -59,7 +69,7 @@ const Historyv1 = () =>  {
 
   useEffect(() => {
     getExerciseActivityByUserId();
-    }, []);
+    }, [reload]);
   
   //get user data
   const getExerciseActivityByUserId = async () => {
@@ -74,6 +84,19 @@ const Historyv1 = () =>  {
     }
   };
 
+  // Delete data
+      const removeData = async (id) => {
+        console.log("id: ", id)
+        // const response = await API.get(`${exerciseActivitiesRoute}/user/${user_id}`, {headers: headers}); // [GET] https://localhost:5000/api/exercise-activities/user/:user_id
+ 
+              const response = await API.delete(`${exerciseActivitiesRoute}/${id}`, {headers: headers})
+          
+              if (response.status === 200) {     // ถ้าลบสำเร็จ 105 จะช่วย reload
+                setReload(!reload);
+                console.log(response);
+              }
+            };
+        
   return (
 
     <ThemeProvider theme={defaultTheme}>
@@ -132,10 +155,20 @@ const Historyv1 = () =>  {
           {/* End hero unit */}
           <Grid container spacing={4}>
             {exerciseActivities.map((exerciseActivity) => (
-              <Grid item key={exerciseActivity} xs={12} sm={6} md={4}>
+              <Grid item key={exerciseActivity._id} xs={12} sm={6} md={4}>
                 <Card
                   sx={{ height: '100%', display: 'flex', flexDirection: 'column' }}
                 >
+                  <CardActions>
+                  <Button size="small">View</Button>           
+                  <Button size="small">Delete</Button>
+
+                  <input type="hidden" id="id" value={exerciseActivity} />
+                  <button onClick={() => removeData(exerciseActivity._id)} className="text-grey pr-5 flex justify-start items-center">
+                  <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 16 16"><path fill="currentColor" fill-rule="evenodd" d="M9 2H7a.5.5 0 0 0-.5.5V3h3v-.5A.5.5 0 0 0 9 2m2 1v-.5a2 2 0 0 0-2-2H7a2 2 0 0 0-2 2V3H2.251a.75.75 0 0 0 0 1.5h.312l.317 7.625A3 3 0 0 0 5.878 15h4.245a3 3 0 0 0 2.997-2.875l.318-7.625h.312a.75.75 0 0 0 0-1.5zm.936 1.5H4.064l.315 7.562A1.5 1.5 0 0 0 5.878 13.5h4.245a1.5 1.5 0 0 0 1.498-1.438zm-6.186 2v5a.75.75 0 0 0 1.5 0v-5a.75.75 0 0 0-1.5 0m3.75-.75a.75.75 0 0 1 .75.75v5a.75.75 0 0 1-1.5 0v-5a.75.75 0 0 1 .75-.75" clip-rule="evenodd"/></svg></button>
+              
+                  <Button size="small">Edit</Button>
+                  </CardActions>
                   <CardMedia
                     component="div"
                     sx={{
@@ -152,22 +185,19 @@ const Historyv1 = () =>  {
                     {exerciseActivity.activity_type_id}
                     </Typography>
                     <Typography>
-                     {exerciseActivity.caption}
+                    {exerciseActivity.caption}
                     </Typography>
                     <Typography>
-                      Descriptons
+                    {exerciseActivity.description}
                     </Typography>
                     <Typography>
                       Calroies
                     </Typography>
                     <Typography>
-                      Date
+                    {exerciseActivity.date}
                     </Typography>
                   </CardContent>
-                  <CardActions>
-                    <Button size="small">View</Button>
-                    <Button size="small">Edit</Button>
-                  </CardActions>
+
                 </Card>
               </Grid>
             ))}
@@ -260,4 +290,4 @@ const Historyv1 = () =>  {
   
 }
 
-export default Historyv1
+export default History
